+++ .changeset/gold-doors-mix.md
@@ -1,0 +1,5 @@
+---
+'@myst-theme/site': patch
+---
+
+Exporting additional components and hooks to help in user web theme development
