+++ packages/site/src/components/Backmatter.tsx
@@ -1,6 +1,6 @@
-import type { GenericNode, GenericParent } from 'myst-common';
+import type { GenericParent } from 'myst-common';
 import { HashLink, MyST } from 'myst-to-react';
-import type { KnownParts } from '../utils.js';
+import { getChildren, type KnownParts } from '../utils.js';
 
 export function BackmatterParts({ parts }: { parts: KnownParts }) {
   return (
@@ -13,21 +13,6 @@ export function BackmatterParts({ parts }: { parts: KnownParts }) {
       />
     </>
   );
-}
-
-/**
- * This returns the contents of a part that we want to render (not the root or block, which are already wrapped)
- * This also fixes a bug that the key is not defined on a block.
- */
-function getChildren(content?: GenericParent): GenericNode | GenericNode[] {
-  if (
-    content?.type === 'root' &&
-    content.children?.length === 1 &&
-    content.children[0].type === 'block'
-  ) {
-    return content.children[0].children as GenericNode[];
-  }
-  return content as GenericNode;
 }
 
 export function Backmatter({
