+++ packages/site/src/components/Navigation/ActionMenu.tsx
@@ -1,0 +1,48 @@
+import { Fragment } from 'react';
+import classNames from 'classnames';
+import { Menu, Transition } from '@headlessui/react';
+import { EllipsisVerticalIcon } from '@heroicons/react/24/solid';
+import type { SiteManifest } from 'myst-config';
+
+export function ActionMenu({ actions }: { actions?: SiteManifest['actions'] }) {
+  if (!actions || actions.length === 0) return null;
+  return (
+    <Menu as="div" className="relative">
+      <div>
+        <Menu.Button className="flex text-sm bg-transparent rounded-full focus:outline-none">
+          <span className="sr-only">Open Menu</span>
+          <div className="flex items-center text-stone-200 hover:text-white">
+            <EllipsisVerticalIcon width="2rem" height="2rem" className="p-1" />
+          </div>
+        </Menu.Button>
+      </div>
+      <Transition
+        as={Fragment}
+        enter="transition ease-out duration-100"
+        enterFrom="transform opacity-0 scale-95"
+        enterTo="transform opacity-100 scale-100"
+        leave="transition ease-in duration-75"
+        leaveFrom="transform opacity-100 scale-100"
+        leaveTo="transform opacity-0 scale-95"
+      >
+        <Menu.Items className="absolute right-0 w-48 py-1 mt-2 origin-top-right bg-white rounded-sm shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none">
+          {actions?.map((action) => (
+            <Menu.Item key={action.url}>
+              {({ active }) => (
+                <a
+                  href={action.url}
+                  className={classNames(
+                    active ? 'bg-gray-100' : '',
+                    'block px-4 py-2 text-sm text-gray-700',
+                  )}
+                >
+                  {action.title}
+                </a>
+              )}
+            </Menu.Item>
+          ))}
+        </Menu.Items>
+      </Transition>
+    </Menu>
+  );
+}
