+++ packages/site/src/components/Navigation/index.tsx
@@ -1,5 +1,7 @@
 export { ThemeButton } from './ThemeButton.js';
-export { TopNav, DEFAULT_NAV_HEIGHT } from './TopNav.js';
+export { TopNav, NavItems, NavItem, DEFAULT_NAV_HEIGHT } from './TopNav.js';
 export { Navigation } from './Navigation.js';
 export { TableOfContents, InlineTableOfContents, useTocHeight } from './TableOfContents.js';
 export { LoadingBar } from './Loading.js';
+export { ActionMenu } from './ActionMenu.js';
+export { HomeLink } from './HomeLink.js';
