+++ packages/site/src/components/index.ts
@@ -1,13 +1,20 @@
-export { ContentBlocks } from './ContentBlocks.js';
-export { DocumentOutline, useOutlineHeight, SupportingDocuments } from './DocumentOutline.js';
-export { FooterLinksBlock } from './FooterLinksBlock.js';
+export { ContentBlocks, Block } from './ContentBlocks.js';
+export {
+  DocumentOutline,
+  useOutlineHeight,
+  SupportingDocuments,
+  useHeaders,
+} from './DocumentOutline.js';
+export { FooterLinksBlock, FooterLink } from './FooterLinksBlock.js';
 export { ContentReload } from './ContentReload.js';
 export { Bibliography } from './Bibliography.js';
 export { Footnotes } from './Footnotes.js';
 export { ArticleHeader } from './Headers.js';
-export { FrontmatterParts, Abstract, Keywords } from './Abstract.js';
+export { Abstract } from './Abstract.js';
+export { Keywords } from './Keywords.js';
+export { FrontmatterParts } from './FrontmatterParts.js';
 export { BackmatterParts, Backmatter } from './Backmatter.js';
 export { ExternalOrInternalLink } from './ExternalOrInternalLink.js';
 export * from './Navigation/index.js';
 export { renderers } from './renderers.js';
-export * from './SkipToArticle.js';
+export { SkipToArticle } from './SkipToArticle.js';
